+++ server/src/main/resources/migrations/V2.2__custom_orders.sql
@@ -1,7 +1,7 @@
 CREATE TABLE CUSTOM_ORDER_ITEM
 (
     id              SERIAL      NOT NULL PRIMARY KEY,
-    client          BIGINT      NOT NULL,
+    client          INT         NOT NULL,
     description     TEXT        NOT NULL,
     date_of_order   TIMESTAMPTZ NOT NULL,
     date_of_receive TIMESTAMPTZ NOT NULL,
